Revert "import inofficial 1.0.4pre by joshua"
This reverts commit 9859b44888b00c0835c0c4997f5f2ae4f6dfaf52.

This change causes 2 bugs with the GBC game Cannon Fodder

( Cannon Fodder (U) (M5) [C][!].gbc
  md5sum 762d6c94874d8ac894ad100c0a4b50ab )

1) the 3D scenes flicker and there are some artifacts in the title
   screen,
2) as soon as mission 2 starts, the CPU goes into an endless loop
   and doesn't process input anymore - this is really nasty, as it
   is even impossible to get out of gnuboy itself, at least in fb
   mode.

i was skeptical about this inofficial patch since the beginning,
since it was not done by laguna himself but by some gamers, and
now it's confirmed that it's broken. afaict this patch includes
3-4 logical changes, if further issues are experienced it may make
sense to try one after the other to see if it fixes them.