add new libao audio backend
2 decades ago, when gnuboy was written, OSS was the main audio backend
for most unices, though it's been replaced with alsa on linux since, so
on a vanilla kernel configuration no sound comes out of the speakers
with gnuboy's default audio backend (except for SDL variant).
unfortunately, using alsa's libasound is quite complex and involves lots
of code.
libao is a portable audio layer, which, on linux, sits on top of alsa
and provides a very simple and nice API, very similar to what OSS used
to be.
ao is now selected by default when --with-sound=yes (default) is passed
to configure, as it's superior to the mostly-defunct OSS backend.