add SDL2 gfx backend
the software scaling code of refresh.c is quite slow and in combination
with low-quality xorg drivers anything > scale 1 caused stuttering in
audio with the existing render backend and unplayable framerate with
scale > 2 in powersave governor.

the new SDL2 backend fixes all of that and even with crappy X11 drivers
that only work well in OpenGL mode scales of 5 and higher run at barely
10% total cpu load.

the SDL2 backend forces the engine to render at scale 1 to a 32bpp
pixel buffer, which is then scaled via hardware; therefore the addition
of delegate_scaling member to fb structure.

unlike the SDL1 backend, this one comes without SDL specific audio backend,
as i assume the SDL2 way to do audio is likely as crappy as the SDL1 way
using async callbacks from a separate thread and other idiocy that requires
about 10x times the amount of code to achieve what the ao backend already
perfectly provides.

so in order to use SDL2 port, use --with-sdl2 and --with-audio=ao.