lcd: prevent from using a scale level that isn't actually supported
when i recently needed gnuboy, rendering was utterly broken and it
require a while to find out it was due to "scale 5" in my gnuboy.rc.
there's not actually any code in gnuboy to support scaling higher
than 4, so the engine just doesn't render anything in the switch
statement and renders the existing junk in the buffers to the screen.

the only backend that supports scale > 4 is the SDL2 one which i
added minutes ago, due to delegating the scaling to the GPU.