undo push-model SDL2 sound backend, use SDL1 approach
while theoretically nicer, the push model causes to get sound
out of sync with the emulator; apparently the emulator needs to
wait until the buffer is actually consumed.
i'll look into fixing that, but until that's done use the SDL1
approach. it's also helpful to have this in commit history
so one can easily go to the SDL1 style sound and compare.