1.0.4-pre backport: fix omission to write RI_SB io reg
this is one of the changed hunks in 1.0.4-pre svn commit
( 9859b44888b00c0835c0c4997f5f2ae4f6dfaf52 ), which was
reverted as one of the changes was bogus (probably the hdma one).

the change looks legit, as it merely adds a missing register
to the write command (it was already in read command).