fix empty pause window in NBA Hoopz
when the pause menu is toggled in NBA Hoopz, a window shifts in
to the right in the lower half of the screen, but it was empty.

this was caused due to WT being calculated based on wrong
assumptions, the assumption that the R_WY register would contain
the line number of the first visible line of the window.

this assumption is wrong, there are various ways to make the
window visible or invisible, one is by using the WIN_ON bit in
LCDC, another is to adjust to WX coordinate to inside or
outside the screen area, and another one is to set the WY
coordinate (assuming in all cases that WIN_ON=1).

in case of NBA hoopz, WY is always 0, as it usually displays
the window on top with the turbo bar and the name of the
currently selected player. WIN_ON is 1 for the entire screen
refresh, and after the top bar has finished painting the
WX coordinate is set to off-screen to make the win disappear.
when the pause menu is on, the WX coordinate is simply shifted
into the visual area on line 100-something.
this made the WT calculation off by > 100 lines and filled
the window with empty sprites happening to be at the mis-
calculated memory region.