fix demo windows in NBA Hoopz
the previous fix wasn't enough for NBA Hoopz:
in "demo" mode, that is when the player does nothing and the game
starts to showcase its capabilities, on top a window of height 8
is displayed, containing the text "DEMO", and when a field goal
happens during the sequence, a second window showing the score is
displayed further down.
using our previous fix, the top line of tiles of the DEMO window
was duplicated inside the score window.
this means the window area contained the top line, followed by
the score window contents.
therefore we need to keep track of how many lines of the window
area have been displayed so far, rather than using the lines
displayed during this window painting period.