implement in-game menu system
if no rom name is passed on the command line, the user is presented
with a menu allowing him to select a rom.
also the esc key is remapped to bring up the menu at any time.

the menu features the following functions:
- rom selection / file browser
- state save/load
- pausing the emulation just by means of bringing it up
- reset
- controller config

a new rc variable "menu" is available to map a key to bring up the
menu, and a "romdir" variable is used to set the initial directory
for the rom browser.

not implemented yet is permanent storage of changed controller config.