change ev_poll() to accept a wait parameter
if provided and set to 1, the backend can use a blocking syscall
to wait without generating cpu usage through polling until an
event happens.

currently only implemented for the SDL2 backend which can make use
of SDL_WaitEvent() - which is currently frustrated by
https://github.com/libsdl-org/SDL/issues/5425 .

this mode will be used by the new menu system to not generate
cpu load while it is open and the user is idle.