loader: fix ub accessing data w/ pointer from before realloc
this caused a crash trying to load a tar.xz that didn't contain
a rom, because the file was a lot bigger than the romsize it
extracted from the "header", so the realloc call resulted in
a new allocation, invalidating the old pointer. that it used
to work in the past with valid roms on the platforms tested
is pure luck.