rom loader: return error rather than dying
it's ok to just die when loading a rom fails for a command line
app, but for an application that's running e.g. on a game console
or handheld with a menu system, it's not really tolerable.

with this change, the frontend has to check for error return code
(i.e. not 0), and then get the error message via loader_get_error()
and decide on whether to abort or just do something else (like
returning to the rom selection menu).