loader: add support for PKZIP format and fix memleaks
until now, only gzipped "zip" files were supported, but not the
pkzip format that's ubiquitously used in the emulator scene.

the pkzip parser is totally simplified and supports only loading
the very first file stream, but this should be sufficient in
99.9% of cases. only few zipped roms contain additional files such
as file_id.diz or readme.txt, those are only compatible if the
.gb/.gbc stream was added as first file.

re: memleaks, until now when the decompressor functions succeeded,
the initial memory copy of the file contents was silently discarded,
and bootrom allocation were never freed either.