add new faster inflate code from miniz
as of upstream commit 5a38f55. this code is a lot faster and makes
zip files on PSP load instantly, while it took several seconds
to load using the old code.
the miniz.h used here was copy/pasted from the upstream one to
include only the necessary bits, and always default to the portable
choice (like in little-endian bitswap macros, and "fast type"
selection).

the old code is still kept to keep supporting .gz files, since
miniz doesn't have code to deal with gzip headers.
we may in the future parse the gzip header ourselves just like
we do in the pkzip case, and switch to the new code for gzip
too.